added -t flag to stop time and memory message from being shown in the… (#358)
Summary:
Adds feature requested in https://github.com/facebook/pyrefly/issues/301. The flag is -t (or --no-timer) and just removes the part of the output that has memory usage and time taken. Simple enough.
To be honest I am not sure how to add it to the printed information when running --help. If shown how to, I can add it and update the commit.

Pull Request resolved: https://github.com/facebook/pyrefly/pull/358

Reviewed By: stroxler

Differential Revision: D75422641

Pulled By: yangdanny97

fbshipit-source-id: 45354f677e2e20d973fd024aa74c1df43a69e4ab