Sign Windows binaries for pyrefly
Summary:
## Background
We are pushing to sign all MSDK builds for Windows targets. We added support
for signing and rolled this out for some use cases:
https://fb.workplace.com/groups/383318912325048/posts/1671551986835061.

See D74278362 for an example of how signing is specified in the config.

## This diff
This diff is generated by a codemod which results in the signing of Windows
binaries for pyrefly.

## What changes can I expect?
 - The Windows binary/ies generated by this MSDK build will now be signed using
   the "in-house" key:
     - Issued to: "Meta Platforms, Inc."
     - Issued by: "DigiCert Trusted G4 Code Signing RSA4096 SHA384 2021 CA1"
 - **Your binary will perform substantially faster! If you have metrics that
   already track this, please ping ejnguyen!**
 - Due to how signing works, the binary itself will be built on a Linux host.

## How can I test this?
The MSDK build should have been triggered already as part of this build (search
for the `MSDK Build Bot` signal). If it hasn't, you can manually trigger
it by leaving `#msdkbuild pyrefly` as a diff comment.

 - Find the `MSDK Build Bot` signal.
 - Validate that the parent job succeeds.
 - Validate that all children jobs succeed as well.

For additional testing, you can download the binaries and actually run them on
your local Windows machine:
 - Find the `MSDK Build Bot` signal. It should have a generated diff attached.
 - On your local Windows machine:
    - `cd C:\open\fbsource`
    - `hg up <generated diff number>`
    - `<path to dotslash shim> <other arguments>`

## What if this results in a broken binary?
If this ends up breaking your binary, please:
 - Revert this diff
 - Once landed, trigger a one-off build:
    - Find your schedule using `bunnylol msdk pyrefly`
    - Click `Trigger Land Bump`.
 - Verify that it generates a working binary (for which the diff should be autolanded).

Note: If the above process _does not_ work, then this diff was not the culprit:
 - Find your schedule using `bunnylol msdk pyrefly`
 - Find the latest diffs that were actually landed in the Diffs column.
 - Revert those diffs.
However, this means that your binary (as built by DotSlash + MSDK) is likely
broken when built from a recent commit! This should be promptly addressed, as
future builds will break your binary (independent of signing)!

## What if I don't want signing?
If you don't want signing (for some reason), do NOT abandon this diff (as
it'll simply get regenerated). Instead, please:
 - Commandeer this diff
 - Change `"sign"` instances from `True` => `False`
 - Publish and land the resulting diff

## I have other questions
Please post here: https://fb.workplace.com/groups/dotslash.users

Reviewed By: yangdanny97

Differential Revision: D75162447

fbshipit-source-id: 163b4631c0873644e621c0225b22a7a144f111b7