Rework bind_attr_assign / bind_subscript_assign
Summary:
It is pretty important for us to be able to handle usages appearing
inside of attribute or subscript assignments for pinning first usage
of definitions that introduce `Var`.

To this end, we need to defer the construction of the `ExprOrValue`
a bit, because once we start tracking usage we need to be able to pass the current
idx to it.

Reviewed By: ndmitchell

Differential Revision: D75422384

fbshipit-source-id: 3d1a1621083e7e3426528d3d7852c290f6d50ac5