Rename the idx of the unpack, add a comment
Summary:
This code is confusing and slightly buggy (it drops annotation information, so
you can assign a type of any value to any variable in an unpacking and Pyrefly
will never complain).

Renaming `key` to something more descriptive helps with following what's
going on here (which is that we make a `Binding::Unpack` for the entire RHS
of this stage of the computation - there might be multiple stages for
multiple layers of nesting - and then we'll use `Binding::UnpackValue` to
extract individual target types from it).

Reviewed By: ndmitchell

Differential Revision: D75422381

fbshipit-source-id: c6617f0e7de83c071d71b38afe06ffa078ef2e93