implement basic completions for module attributes
Summary:
this diff implements basic completions for the `y` in `from x import y`.

we do this by reusing the `import_at` api implemented in the last diff. unfortunately, there are a few gaps:
- `from x import <>` does not work
- `from x imp<>` unfortunately works (because ruff's parser makes that an ImportFrom node with module = `x` and names = ['imp'])

I will fix these in follow-up diffs, but I believe the current implementation is better than nothing for now

Reviewed By: SamChou19815

Differential Revision: D75477356

fbshipit-source-id: 5fabe76f474a842f2d5e8319b50c7cc567de455d