represent imports better
Summary:
by using `Ast::imports` to find imports, we considered the entire import stmt including keywords in what we consider a "module". But in reality, we should only consider the names*.

This diff makes a `ImportIdentifier` enum, walking the statements and grabbing the necessary items.

For LSP features like completion, I believe the x in `from x import y` is very similar to the the x in `import x`. For that reason I make both of these `x`s `ImportIdentifier::Module`. But the names are different and get their own enum variant (to be used in next diff).

Reviewed By: SamChou19815

Differential Revision: D75473401

fbshipit-source-id: ddcf1b5da352e7bb9ab59e046784f09587f1cd02