Report path validation errors at severity WARN
Summary:
We used to report these as warnings (reasoning in D72078439), but I changed them to errors in the course of fixing an issue where warnings would print at inconvenient times and mess up our progress bar.
This diff switches the errors back to warnings, but this time in a way that doesn't mess up the progress bar.

Reviewed By: grievejia

Differential Revision: D75484253

fbshipit-source-id: e848f1327d489c24df5a092844432e3ead0220b4