Break off aug assign [1/N]
Summary:
I realized that the use of a single god function for augmented assignments and normal
assignments is not only causing me migraines, but is also semantically wrong: we're creating a
`Key::Definition` here but this is not quite right, an augmented assignment is not really a
definition.

--------

This shows up as user-facing if you do several augmented assignments to an undefined variable;
in reality that variable doesn't exist so every use should be an error (it's equivalent to
calling `print(variable_does_not_exist)` repeatedly on the same variable), but we think the
name is defined after the first augmented assign.

----------

This is a super minor issue, and if it weren't for BE considerations and
first-usage tracking (see below) I'm not sure it would motivate the change, but I still
think our behavior is probaby incorrect; Pyright flags every repeated aug_assign of an undefined
variable as a type error which feels more right to me.

More importantly, first-usage tracking for determinism will be very difficult if we don't do a
better job of isolating the `ensure_name` happening in the `Expr::Name` case - eventually that
ensure has to happen with knowledge of the `Key` whose binding we are currently computing, or
else we would fail to detect that in
```
my_list = []
my_list += [5]
```
the `+=` is a valid first-usage and should pin the type of `my_list`.

Given that the augmented assignment is further complicating code that's already extremely
hard to reason about, the fact that there are semantic issues feels like reason enough to
just fork the handling.

------

This diff does a naive fork of our target-binding god-function and a tiny bit of simplification
since we no longer need an `is_aug_assign`. I'll inline some code and continue simplifying next.

Reviewed By: rchen152

Differential Revision: D75471650

fbshipit-source-id: 7405118ea8502cc66f8f687ffc4a729b470b2d24