Remove unneeded TODOs
Summary:
Now that I can see what's going on better, it's pretty clear that
the TODO I had about contextual typing is not relevant: an augmented
assign is a dunder operation plus an assign, and since dunder operations
are calls this means that any contextual typing happens in the
dunder operation rather than the assignment phase.

So get rid of those todos.

Reviewed By: ndmitchell

Differential Revision: D75527882

fbshipit-source-id: 0475569511180725e19d441a802f600e638ff552