Improve inputs to check_second_arg_is_class_object
Summary: I noticed while reviewing a PR that we make some (pre-existing) questionable assumptions in check_second_arg_is_class_object and check_type_is_class_object that we're dealing with a function named `isinstance`, when it could be `issubclass`. As the first step to fixing this, pass in a FunctionKind to check_second_arg_is_class_object.

Reviewed By: grievejia

Differential Revision: D75551332

fbshipit-source-id: 2631ee7a897b2de58e1b2c45f6b82f970bb0df80