Fix async return type wrapping logic
Summary:
For normal async functions, we always wrap its return type with a coroutine. But for async generators, we don't really want to do the wrapping. The difference between normal async funcs and async generators is that the latter contain at least one `yield` statement syntactically, whereas the former do not. E.g.
```
async def foo() -> int:
  return 42
reveal_type(foo())  # Reveals Callable[[], Coroutine[Any, Any, int]] (with wrapping)

async def foo() -> AsyncGenerator[int, None]:
  yield 42
reveal_type(foo())  # Reveals Callable[[], AsyncGenerator[int, None]] (without wrapping)
```

We used to unconditionally wrap all async return types with `Coroutine`. D68927030 made the correct observation that we should stop doing that for async generators, but the way it checks for async generator is incorrect -- we are supposed to match CPython runtime and perform a syntactical scan over the function body looking for `yield` statement, but that diff did a type-level subset_eq check with `AsyncGenerator` instead. See the linked github issue on why the type-level check is problematic -- e.g. if the return type of an async is actually a gradual type like `Any | None`, we'll incorrectly conclude that the function is an `AsyncGenerator` (because the subtyping query always succeeds with the `Any`) and refuse to do the wrapping!

Later came D69713115, which correctly implements the syntactical check when solving for `Binding::ReturnType`. However, that diff did not undo D68927030, stilling relying on D68927030 to do the coroutine wrapping in the case where an explicit return annotation is provided, and at the same time duplicated the coroutine wrapping in the case where an explicit return annotation is not provided.

Then we have D74420059, which observed the duplicated wrapping in the case where an explicit return annotation is not provided. But all it did was to merely remove the wrapping in `Binding::ReturnType`, still leaving the incorrect wrapping logic of D68927030 intact!

This diff proposes that we should really address the root cause, by removing the async generator subtyping check introduced in D68927030, as well as addressing 2 direct consequences of the removal:
- Now that we no longer have the D68927030 wrapping, the duplicated wrapping issue observed by D74420059 no longer applies. We should simply back out that diff.
- With the D68927030 wrapping gone, we revealed the issue in D69713115 where it forgets to do the wrapping when explicit return annotation is provided. Filling in that missing part makes all behaviors sensible again.

With this change, the solving logic for `Binding::ReturnType` is exclusively responsible for any kind of return annotation wrapping.

This diff should fix https://github.com/facebook/pyrefly/issues/392

Reviewed By: ndmitchell, stroxler

Differential Revision: D75573622

fbshipit-source-id: 2df894374653f8a139ba21a8eeb4789f2721df07