Make ignore comments apply to the line that comes after further comment block
Summary:
ignoreignore

When an ignore comment is followed by more comments, e.g.
```
# pyrefly: ignore
# This is a legit error but we cannot fix it for reasons X, Y, and Z
x: int = derp()
```
We currently count that ignore as not applicable to the error, which is sad because it precludes the possibility of adding more explanatory comments on the ignore.

This diff attempts to relax the behavior a bit, by "extending" the type error range and have them include more comment lines above, before performing the ignore coverage check.

Reviewed By: stroxler

Differential Revision: D75561541

fbshipit-source-id: 4127f8bb041133d47a7d991c685ccef865945f18