Decouple augmented assignment from `bind_assign`
Summary:
Augmented assignment is pretty weird, in ways that become more evident when we
want to do usage tracking. This diff decouples it from the normal assignment
helper; I retain the existing behavior (for now) but add some comments where
the optimal semantics are unclear.

It behaves a bit like a dunder call plus an assignment, which means that:
- it arguably doesn't actually *define* the name being assigned; this is
  debateable and I'm not making changes, but I left a TODO because this
  might affect IDE behavior.
- it arguably shouldn't change the `FlowStyle`, which it currently does
  because `bind_key` always sets a flow style. I don't actually know
  whether this matters and left another TODO
- it depends, for evaluation of the synthetic dunder call, on the LHS.
  This is the actual critical difference: it's not possible (without
  making the normal `bind_assign` more complex than I think is worthwhile
  for a single caller) to use the normal first-usage tracking logic
  because we have to actually treat the LHS as a usage, which no other
  scenario wants to do.

The third bullet is my primary motivation for decoupling them, but I do think
the TODOs are also worth revisiting at some point.

Note that the attribute and subscript cases are fundamentally different, they
actually *do* behave from both a flow-info-tracking and usage-tracking point
of view much more like a "normal" assignment to an attribute and subscript
(because even the normal case depends on the LHS for those). For this reason,
I left them mostly unchanged, other than now we can define `make_assigned_value`
directly.

Reviewed By: ndmitchell

Differential Revision: D75560166

fbshipit-source-id: d73212f84974319f897b0c76bbb019270e595ff8