Tweak for loop code ordering
Summary:
In our fix for the case where a variable appearing as part of the
iterator would wind up with a var forced to Unknown for
https://github.com/facebook/pyrefly/commit/3b74aface52a46dc2a62add74a17c79c835ddd42,
we fixed the bug but left the (rust) code in an ordering
that is ripe for confusion, because we fixed it by moving only
the `ensure_expr` and not the actual bind operations to be
before `setup_loop`.

Trying to manage ensures in a way that is mandatory for usage tracking
can easily re-introduce the bug, so I'm shifting the order beforehand
to keep the tricky diffs as clean as possible.

Reviewed By: ndmitchell

Differential Revision: D75583320

fbshipit-source-id: 3e87f992250005ca9910e72cb844b6d1777f4233