Regression tests for async functions that are not generators
Summary:
In D75573622 we fixed a major bug in async function wrapping.

But we didn't actually test it - we added an additional snippet of code to an existing test
case, but that snippet was unaffected by the fix (it worked fine on trunk).

This diff backs out the unnecessary test change, and adds a new test that does exercise
the bug fix (I verified that both of the assertions fail prior to D75573622)

Reviewed By: grievejia

Differential Revision: D75611905

fbshipit-source-id: 36667fca2eab4984c324976ed7c1ca61d0804746