Add the plumbing needed to cache legacy variance results and only retrieve them where needed
Summary:
Instead of reading variance results directly from type parameters when we need them for the subset check, we cache those results via a binding and retrieve them from the map directly.

We cache the results via a binding where we collect the variance information from the type parameters and store them in a map.

My idea is that the next step should be extending this map to include scoped type variables' inferred variances.  We also need to extend the new variance binding to include additional class information.

Reviewed By: stroxler, rchen152

Differential Revision: D75541737

fbshipit-source-id: a1fe52ad73814c37fcb93b727e341d3f83cd494e