Add some comments on loop headers
Summary:
After having a bug related to when loop headers are and are not supposed
to be bound inside of a loop's flow environment (vs the flow environment
just before the loop), I figured this is tricky enough to deserve some
actual discussion.

The `while` loop is actually the trickier one; because we need the narrow
ops *before* we set up the loop, the (rust) code ordering is a little
odd looking, but it's actually essential that we delay the `ensure` until
after the loop is set up.

Reviewed By: ndmitchell

Differential Revision: D75583323

fbshipit-source-id: 1baf86ef5cd3b3ef9dc9460fadf2a7d0ed0b445d