Fully resolve comprehension iterators before looking at targets
Summary:
The following snippet is legal:
```
x = [1, 2, 3]
y = [x for x in x]
assert_type(y, list[int])
```

Today, we handle this by ensuring the iterator (the rightmost `x`
above, after the `in`) prior to adding the lvalue (the middle `x`)
to our scope. But this doesn't play well with usage tracking, because
at the point when we're fiddling with the order of `ensure` vs
`expr_lvalue` we don't yet have access to the idx of the binding we are
about to compute, and the ensure needs to know that.

Inspired by a suggestion from Jia that this really should behave sort of
like defining and then immediately calling a function, I decided the best way
to handle this is probably to just resolve the `IterableValue` *fully* before
we start looking at the target, putting it in a `Key::Anon`, and then use
indirection via a `Forward` to pass that to the target-handling code.

It looks at first glance like this costs us some contextual typing, but that's
a bit misleading: you can't annotate a comprehension target anyway, so
there was never a scenario where contextual typing actually mattered for
comprehensions (this isn't true in other contexts such as for loops)

Reviewed By: grievejia

Differential Revision: D75608507

fbshipit-source-id: ab6411b721119f1715296ce8f289a62b36c90fc5