Target binding manages usage / ensure [5/N]: bind_target_impl
Summary:
**About this stack**

*Requirements at a high level*

In this stack, I eliminate the responsibility of ensuring normal expressions
from callers of `target.rs` code.

Instead of directly ensuring terms they depend on, callers will pass terms
(as `&mut Expr` *into* the `target.rs` logic, which will handle ensuring them

*Requirements: the wierder bits*

We have to deal with a lot of complications in our solution:
- Some callers don't *have* any terms - for example there are a couple places we just produce a `Binding::Type(Any)` out of thin air. So we need some way to still allow callers to skip the expression logic.
- Some callers shouldn't have expressions ensured because they can't use cookie-cutter logic, e.g. type var assignment. Since it turns out *all* such callers are type vars and we probably don't need to pin first-usage on type vars anyway, they can just use the approach above after ensuring the expressions themselves (with no usage tracking)
- Python has multi-target assignments, like `x = y = foo()`. This means that *internal* to `target.rs`, we'll have to have a mechanism to explicitly turn the ensure logic *off*. This won't mess up first-usage tracking because the first usage will just be the first target, but it does complicate the internals.

*The approach we'll take*

All `_impl` functions in `target.rs` will be private, and they will generally include several knobs on the functionality:
- take an Option<&mut Expr> to ensure and pass to callbacks
- take a callback that *accepts* an `Option<&Expr>` to construct bindings or in some cases `BindingOrExpr` values
- take a flag `ensure_assigned` to turn *off* the ensure.

In all callsites (which are internal to `target.rs`), we will:
- either pass `Some(expr)` and then in the callback `.unwrap` the `&Expr`, or
- pass `None` and ignore the

In addition, the callers passing `None` will always set `ensure_assigned` to `false`, although this is just an optimization (there would be nothing to ensure anyway).

Mutually recursive calls will always propagate `ensure_assigned`, and the mutli-target assignment handler will make sure to pass it as `true` only once.

Someone just playing type tetris might be tempted to think the flag is redundant, but this is not possible: we still need to be able to *pass* an `Expr` through (possibly mutually-recursive unpackings involved in) mutli-target assigns, even though we can't double-ensure. So it really is necessary

**This diff**

Closes the loop on the implementation by applying the plan to `bind_target_impl`; now all the mutually recursive functions behave as expected.

We still haven't changed the public functions that call into these yet though (which also means most callsites haven't been adapted yet)

Reviewed By: ndmitchell

Differential Revision: D75583317

fbshipit-source-id: 2c784bef01d14f2636419722e08b30aa05c45dca