Refactor attribute and subscript assign
Summary:
Because of assignment narrowing for subscripts and attributes, the
subscript and attribute target handlers have two places where they
do things differently depending whether the base is an identifier:
- a branch for the key/idx:
  - it's anonymous for no narrowing, we just check the assignment
  - but it's a PropertyAssign when we assign to a name, b/c we need to narrow
- another branch for the flow info
  - there's nothing to do in the no-narrowing case
  - but in the narrowing case we need to update the flow info

We have been duplicating the entire logic up until now, which seemed worth
it because the flow is simpler to read with a single branch than two branches
when there's not much code involved.

But the previous diff moved the `ensure` operations to live after we define the
`idx`, which shifted the balance in my opinion to where the duplication is no
longer a good idea, it would be too easy to have a bug in one branch and not
the other.

Refactoring so that we just branch on the `idx` and the flow updates, and the
rest of the logic is a single analysis.

Reviewed By: yangdanny97

Differential Revision: D75618676

fbshipit-source-id: 350caa5784d1e1af202070439ff69fc03f2ab473