Prepare `bind_{attr,subscript}_impl` for usage tracking
Summary:
Ensures need to happen after we have the idx.

This is a little messy, the code was already somewhat duplicative between
the has-identifier-base vs not cases (these cases differ because narrowing only
applies when there's an identifier base) and now it is a bit more so.

I'll try to refactor, but I figured it's better to start with something
simple and correct first.

Reviewed By: yangdanny97

Differential Revision: D75618677

fbshipit-source-id: d63fafbad0760b776b03e56f5aa25d1de75fa63b