Get rid of early return in `ensure_expr`
Summary:
I'm generally mildly opposed to the use of early returns in very large
match statements, they tend to be hard to follow and a bit brittle when
we need refactors. I realized that sometimes it's a good approach, but
in my mind the benefits need to be pretty high.

In this case, at one point we were really doing a lot of stuff in the
after-match block, but we already gradually eliminated most of them
(most recently, after some scoping bugs forced comprehensions to not fall through).

At this point there are only three cases (Name, Yield, and YieldFrom) that *do*
fall through, and one of them (Name) is a red herring because it has no sub
expressions!

So we now have about 10 early `return` statements to save us from 2 extra
`recurse_mut` calls on yield and yield from. This is a bad tradeoff already.

Also, yield and yield from both get their own bindings, which means for
usage-tracking and determinism purposes I'm going to have to push the ensure down
soon anyway, which literally means that the only case falling through would be the
`_` case.

Going back to my point about not liking early returns, I would like to point
out that by the time I implement usage tracking, *every single use* of fall-through
was either an outright bug that we've fixed, or a case where evolving
requirements will have forced us to update. I've observed this pattern several times
on large matches using early returns, which is where my instinct that this tends
to be a code smell comes from.

Reviewed By: grievejia

Differential Revision: D75701660

fbshipit-source-id: 03b4cf8a073519f5e18a775e547efb55aa05cefd