Clean up some flow-merging cruft
Summary:
I realized that `merge_flow` is not used anywhere outside `bindings`, so it
can be made private. But at that point, it doesn't feel worth the indirection
between `merge_flow` and `merge_flow_is_loop` so I just unify them under
the name `merge_flow` (which I think makes more sense to read - `is_loop` is
a *flag*, the function merges flows whether or not it's a loop. When inlay
hints are on, this naming reads very clearly).

Reviewed By: grievejia

Differential Revision: D75701659

fbshipit-source-id: baf250b3091dbd410847274971203195015d7532