Use a consistent key for returns
Summary:
We are currently using a `Key::Anon` for illegal top-level
returns, and a `Key::ReturnExplicit` for legal ones.

This would be hard to preserve in the refactors I need for
usage tracking, but it actually makes no difference.

Putting up a trivial diff to change it so the more substantial
refactors can be pure refactors.

Reviewed By: grievejia

Differential Revision: D75708640

fbshipit-source-id: 218ed3f1151365a020f212ebe605c80dd60b18b8