Use `insert_binding` for `bind_definition` and add a new helper
Summary:
I'm about to need a version of `bind_definition` that lets you use a preexisting
idx (because a lot of places need to start creating the idx ahead of time to allow
usage tracking in `ensure`).

While I was looking at that, I noticed we have a `self.table.insert`. This isn't
wrong, but I'd rather use the more common and public `self.insert_binding` when
possible, just for consistency.

Reviewed By: grievejia

Differential Revision: D75714593

fbshipit-source-id: 7e7595e44e1a98f2b531457c2a9bbab9b5f1c21a