Push all of the synthesized class def logic to `class.rs`
Summary:
For usage tracking, I need to rearrange this code slightly so that we can
create the class indices prior to ensuring expressions.

The first step is to get all the logic in one place; we don't really need
all these `assign` helpers, two layers (the use-case-specific functions
and then the general synthesized class def handler) are enough.

Reviewed By: grievejia

Differential Revision: D75714594

fbshipit-source-id: f76371543ff89d536c35be9bb2ba7a06158bd84e