Simplify yield handling by separating yield vs yield_from
Summary:
Smashing these together into an Either was making the code pretty complex,
both the type itself and the code using it (which constantly has to match)
are pretty boilerplate-y.

It's an easy refactor to just separate them, which will make future
refactors and also just skimming the code easier.

Reviewed By: grievejia

Differential Revision: D75701663

fbshipit-source-id: 4c2275dc894e8c54e1f57f83ca98a389e0ed4e2a