Push the `class_indices` creation up for synthesized classes
Summary:
The expressions we analyze in synthesized class definitions are all ensured
in the use-case-specific helpers.

This means we want to create the indices (in particular, the idx_class_object
index which tells us the `Key` / `Binding` pair of the final class object bound
in the scope that creates a synthetic class) at the top of those use-case
specific helpers so that we can add usage tracking to our `ensure` calls.

Reviewed By: grievejia

Differential Revision: D75714595

fbshipit-source-id: 0c551c2c250070989811b0fc358e3482d2ff5884