Stop counting `x` in `local x` or `global x` as definitions
Summary:
The `global` and `nonlocal` statement only declares the existence of a name in the current static scope. They would never "define" a name -- the name is always defined somewhere up the scope stack. It's incorrect to include them in definition collection logic.

This simple change surprisingly fixes 2 unit tests.

Reviewed By: stroxler

Differential Revision: D75729810

fbshipit-source-id: afa8a75eed6cdc3f9bed4dacb0905f3253a7a5e2