Prepare function.rs for usage tracking (behavior change!)
Summary:
**About this diff, in the context of determinism work**

My original intent with this diff was just to define a couple of `Idx<Key>`
values earlier so it will be easy to pass them to `ensure` when we turn on
usage tracking.

Now, if Pyrefly were well-behaved in the sense that a change to the ordering
of keys cannot affect solve results in a single thread, I would have expected this
to make no difference - it's pretty trivial to read off from the diff that we're not
actually doing anything different here, all we do is create the `idx` we would
need before ensuring expressions (which does not impact lookups - the scope
isn't modified until `bind_definition_idx`!)

**The unexpected result**

For reasons that are slightly mysterious to me, defining the `Idx<Key>`
earlier is actually causing behavior changes.

Here's what I think is happening, I suspect it's actually closely related
to nondeterminism:
- Before this change, the location of the function key in our table was *after*
  the function body. That meant that we would wind up hitting last-statement
  bindings (used in implicit return analysis) *before* we analyzed the function
  name itself. I suspect that there's some cycle where that analysis behaved
  in a way that depended on getting a recursive Var back for the function itself;
  not coincidentally at least one of the examples is a recursive function, and
  another one involves two defs of the same name.
- After this change, the function idx is now first. That means if there's some
  cycle, the order in which we hit recursion is different.

The fact that there's a change here at all is still a bit fishy - my understanding
is that (without any formal reasoning actually proving this should be the case)
we more or less believed that our recursion and cycle resolution logic is invariant
to the order in which one enters the cycle. But given that this change is pretty
trivial other than potentially impacting that order, I would argue this is some
strong evidence that we actually do have some entry-point-sensitive logic.

It's at least plausible that we won't be able to get 100% determinism without
tracking this down.

That said, I'm not inclined to hold up my current effort: this diff trades
several mysterious bugs for a new mysterious bug where `super()` is behaving
as if it were in an instance method, when we are not in an instance method.
This seems like an acceptable tradeoff given the urgency on determinism, and
debugging the new bugs (which again, are in exchange for fixing other bugs
anyway!) seems likely to teach us something new about whatever is causing
these odd method-scope issues.

Reviewed By: ndmitchell

Differential Revision: D75722156

fbshipit-source-id: 94da553a3ebdcc9678e3db4e2398723867187d75