Introduce Usage to the ensure API
Summary:
In order to do usage tracking, we need a clean way of handling the
95% case where all name lookups need to be usage-tracked, but also
deal with the remaining cases where I really don't think we want
usage tracking (static types and pattern matching).

I considered a few ideas that I decided are asking for trouble:
- Making the API take an `Option<Idx<Key>>`. This seems like a bad idea
  because it becomes very easy to "cheat" by passing `None`, and there's
  not even a clear marker in the code that we're doing it (nothing easy
  to ripgrep for, for example). Given that I fully expect and even *want*
  people who don't understand the system end-to-end to be able to hack on
  it, this seems like pretty clearly a no-go.
- Making two versions of the API, a usage tracked one and a non-usage-tracked
  one. This is a bit ugly because it would require an `impl` that takes an option
  and then two public shims... we do this sort of thing in a few places where
  it's really necessary (e.g. `target.rs`) but it makes the code harder to skim.
  In addition, it *still* feels to easy to mess up:
  - An onboarding developer who's just hacking away would be tempted to use the
    untracked API, and it might slip through code review
  - Inside of `expr.rs`, we would *have* to be still passing around
    `Option<Idx<Key>>`, which means that the disadvantages of the first bullet
    (where it would be way too easy to pass `None` by accident) apply in full
    force.

The idea I came up with is to have a `Usage` datatype that is effectively
an `Option<Idx<Key>>`, but has its own type identity. Callers will *have* to
give us a `Usage`, so if they want no tracking they'll have to use
`Usage::NoUsageTracking`. This is easy to search for, and lends itself to
a comment and even a clear variable name (I call it `no_usage` in the first use
case, inside `ensure_type`).

----

An added bonus of this approach is that it allows me to put the usage tracking
plubming in place without figuring out all the details in one diff, because
I am able to make a placeholder `NotImplemented` variant; I'll be spending
the next several diffs cutting all of these over until I can eliminate it.

Reviewed By: ndmitchell

Differential Revision: D75742941

fbshipit-source-id: bc90a99d042cbc878ad49fba644b77bb35dbd0c0