Untroduce `User`, convert `pattern.rs` to usage tracking, introduce bindings helpers
Summary:
The previous diff introduced usage tracking plubming to the ensure API.

This diff wires up our first module to actually use it:
- Pattern.rs has no usage tracking on match patterns
- The match subject and guard do need usage tracking

-----------

I was originally expecting to just directly use `Usage` for this, but
I realized it is quite a bit more type safe to create a `User` abstraction,
which gives us not one but two benefits:
- First off, to have an ergonomic binding API I wanted to make it easy
  to pass the "identity of a user", and allowing passing a `Usage` directly
  would require a panic when it isn't `Usage::Idx`
- Second, as an extra bonus we can make binding *consume* the `User`, which
  is great because then the user goes out of scope in calling code, and we
  are protected from trying to double-bind it (which would panic, but now
  we'll get type errors in the IDE) or trying to convert it to a "usage" after
  it is already bound (which technically *could* be okay, but is a code
  smell and I think banning it will make the semantics of the analysis
  much clearer).

I'm pretty sure these protections are going to help me not make any errors
in this stack, as well as helping future maintainers.

--------------

Just a naming note: I think `User` / `Usage` might be a little awkward, it feels
odd to call an idx a "user" to me.

I really just needed to decide on something and start cutting things over to
usage tracking without obsessing about a name. I'm happy to rename
things at the top of the stack; maybe `s/User/Usage/` and `s/Usage/Use`
would be better? Open to suggestions. Another possibility would be
to just rename `User` to `Current` or `InProgres`, which kind of describes
the role it plays as you skim the code - it's a wrapper around the key of the
currently-in-progress binding

Reviewed By: grievejia

Differential Revision: D75742940

fbshipit-source-id: 76d5dbccebbd50773711f756968df8a618864f54